requirements.txt
git+https://github.com/DeepCube-org/dynamicworld.git#egg=dynamicworldWrap
numpy==1.26.4
pytest==8.1.1
python-dotenv==1.0.1
rasterio==1.3.9
s2cloudless==1.7.2
sentinelhub==3.10.1